requirements.txt
opencv-python
streamlit 
pillow 
easyocr